+++ scanner/src/test/java/MinionProcessorTest.java
@@ -18,6 +18,7 @@ public class MinionProcessorTest extends AbstractProcessorTest {
 
   @Override
   protected NotificationDto process(File directory) throws IOException {
+    instance.setRootPath(directory.toPath().getParent());
     return instance.process(directory, TimeZone.getTimeZone("America/Toronto"));
   }